src/requirements.txt
#certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
opencv-python==4.6.0.66
Pillow==9.2.0
PyMySQL==1.0.2
python-dotenv==0.20.0
requests==2.28.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
gunicorn==20.1.0
dataclasses==0.6
html-table-parser-python3==0.2.0
six==1.16.0
typing_extensions==4.1.1
python-dateutil==2.8.2
pandas
